+++ api/src/main/java/org/allaymc/api/entity/component/EntityBaseComponent.java
@@ -577,11 +577,14 @@ public interface EntityBaseComponent extends EntityComponent, CommandSender, Has
     }
 
     /**
-     * Get the base offset of this entity.
+     * Get the network offset of this entity.
+     * <p>
+     * The network offset is the additional offset in y coordinate when sent over network.
+     * This is mostly the case for older entities such as players and TNT.
      *
      * @return the base offset of this entity.
      */
-    default float getBaseOffset() {
+    default float getNetworkOffset() {
         return 0f;
     }
 
@@ -610,6 +613,31 @@ public interface EntityBaseComponent extends EntityComponent, CommandSender, Has
      */
     default float getGravity() {
         return 0.08f;
+    }
+
+    /**
+     * Get the drag factor when on ground of this entity.
+     * <p>
+     * This factor will be multiplied to the motion along x and z axis every tick.
+     * The bigger the factor is, the quicker the entity will stop in x-axis and z-axis.
+     *
+     * @return the drag factor when on ground of this entity.
+     */
+    default float getDragFactorOnGround() {
+        return 0.1f;
+    }
+
+    /**
+     * Get the drag factor when in air of this entity.
+     * <p>
+     * This value is similar to {@link #getDragFactorOnGround}, however this value
+     * will be used to reduce motion along x-axis and z-axis when the entity is not on ground.
+     * This value will always be used for motion along y-axis.
+     *
+     * @return the drag factor when in air of this entity.
+     */
+    default float getDragFactorInAir() {
+        return 0.02f;
     }
 
     /**
@@ -725,32 +753,35 @@ public interface EntityBaseComponent extends EntityComponent, CommandSender, Has
     }
 
     /**
-     * Knockback the entity.
-     *
-     * @param source the source of the knockback.
+     * @see #knockback(Vector3fc, float, boolean, float)
      */
     default void knockback(Vector3fc source) {
         knockback(source, DEFAULT_KNOCKBACK);
     }
 
     /**
-     * Knockback the entity.
-     *
-     * @param source the source of the knockback.
-     * @param kb     the knockback strength to apply.
+     * @see #knockback(Vector3fc, float, boolean, float)
      */
     default void knockback(Vector3fc source, float kb) {
         knockback(source, kb, false);
     }
 
     /**
-     * Knockback the entity.
+     * @see #knockback(Vector3fc, float, boolean, float)
+     */
+    default void knockback(Vector3fc source, float kb, boolean ignoreKnockbackResistance) {
+        knockback(source, kb, ignoreKnockbackResistance, kb);
+    }
+
+    /**
+     * Knockback the entity with specified kb value.
      *
      * @param source                    the source of the knockback.
      * @param kb                        the knockback strength to apply.
      * @param ignoreKnockbackResistance {@code true} if the knockback resistance should be ignored.
+     * @param kby                       the knockback strength in y-axis.
      */
-    void knockback(Vector3fc source, float kb, boolean ignoreKnockbackResistance);
+    void knockback(Vector3fc source, float kb, boolean ignoreKnockbackResistance, float kby);
 
     /**
      * Apply the entity event to the entity.
@@ -836,7 +867,7 @@ public interface EntityBaseComponent extends EntityComponent, CommandSender, Has
     /**
      * Check if the entity's eyes is in water.
      *
-     * @return {@code true} if the entity is in water, otherwise {@code false}.
+     * @return {@code true} if the entity's eyes is in water, otherwise {@code false}.
      */
     default boolean isEyesInWater() {
         var dim = getDimension();
@@ -845,6 +876,20 @@ public interface EntityBaseComponent extends EntityComponent, CommandSender, Has
 
         return eyesBlockState.getBlockType().hasBlockTag(BlockTags.WATER) &&
                eyesBlockState.getBlockStateData().computeOffsetShape(MathUtils.floor(eyeLoc)).intersectsPoint(eyeLoc);
+    }
+
+    /**
+     * Check if the entity is touching water.
+     *
+     * @return {@code true} if the entity is touching water, otherwise {@code false}.
+     */
+    default boolean isTouchingWater() {
+        var dim = getDimension();
+        var loc = getLocation();
+        var blockState = dim.getBlockState(loc);
+
+        return blockState.getBlockType().hasBlockTag(BlockTags.WATER) &&
+               blockState.getBlockStateData().computeOffsetShape(MathUtils.floor(loc)).intersectsPoint(loc);
     }
 
     /**