+++ api/src/main/java/org/allaymc/api/entity/interfaces/EntityTnt.java
@@ -1,7 +1,8 @@
 package org.allaymc.api.entity.interfaces;
 
 import org.allaymc.api.entity.Entity;
+import org.allaymc.api.entity.component.EntityTntBaseComponent;
 
-public interface EntityTnt extends Entity {
+public interface EntityTnt extends Entity, EntityTntBaseComponent {
 
 }