+++ server/src/main/java/org/allaymc/server/entity/impl/EntityFallingBlockImpl.java
@@ -13,7 +13,6 @@ public class EntityFallingBlockImpl extends EntityImpl implements EntityFallingB
     public EntityFallingBlockImpl(EntityInitInfo initInfo,
                                   List<ComponentProvider<? extends Component>> componentProviders) {
         super(initInfo, componentProviders);
-        ;
     }
 
     @Delegate