+++ server/src/main/java/org/allaymc/server/world/service/AllayEntityPhysicsService.java
@@ -37,6 +37,7 @@ import org.joml.primitives.AABBf;
 import org.joml.primitives.AABBfc;
 
 import java.util.*;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentLinkedQueue;
 
 import static java.lang.Math.*;
@@ -60,9 +61,6 @@ public class AllayEntityPhysicsService implements EntityPhysicsService {
     private static final float STEPPING_OFFSET = 0.05f;
     private static final float FAT_AABB_MARGIN = 0.0005f;
     private static final float MOMENTUM_FACTOR = 0.91f;
-    private static final float GROUND_VELOCITY_FACTOR = 0.1f;
-    private static final float AIR_VELOCITY_FACTOR = 0.02f;
-    private static final float DRAG_FACTOR = 0.98f;
 
     private static final float WATER_FLOW_MOTION = 0.014f;
     private static final float LAVA_FLOW_MOTION = 0.002333333f;
@@ -134,9 +132,20 @@ public class AllayEntityPhysicsService implements EntityPhysicsService {
 
     protected void cacheEntityCollisionResult() {
         entityCollisionCache.clear();
-        entities.values().forEach(entity -> {
+        var map = new ConcurrentHashMap<Entity, List<Entity>>();
+        // Compute colliding entities in parallel, because computeCollidingEntities()
+        // will be an expensive method if there are a lot of entities. Method
+        // computeCollidingEntities() should be safe to call in parallel
+        entities.values().parallelStream().forEach(entity -> {
             var collidedEntities = computeCollidingEntities(entity, true);
-            if (collidedEntities.isEmpty()) return;
+            if (collidedEntities.isEmpty()) {
+                return;
+            }
+            map.put(entity, collidedEntities);
+        });
+        map.forEach((entity, collidedEntities) -> {
+            // These two operations is not thread-safe, so simply do them synchronously
+            // as the two operations shouldn't be slow
             entityCollisionCache.put(entity.getRuntimeId(), collidedEntities);
             collidedEntities.forEach(entity::onCollideWith);
         });
@@ -291,10 +300,14 @@ public class AllayEntityPhysicsService implements EntityPhysicsService {
 
         var finalMotion = new Vector3f();
         if (hasWaterMotion) {
-            finalMotion.add(waterMotion.normalize().mul(WATER_FLOW_MOTION));
+            // Multiple water flow vector may cancel each other out and let the final motion
+            // result in zero vector, so we still need to use normalizeIfNotZero() here to
+            // prevent NaN
+            finalMotion.add(MathUtils.normalizeIfNotZero(waterMotion).mul(WATER_FLOW_MOTION));
         }
         if (hasLavaMotion) {
-            finalMotion.add(lavaMotion.normalize().mul(dimension.getDimensionInfo() == DimensionInfo.NETHER ? LAVA_FLOW_MOTION_IN_NETHER : LAVA_FLOW_MOTION));
+            // Same to above
+            finalMotion.add(MathUtils.normalizeIfNotZero(lavaMotion).mul(dimension.getDimensionInfo() == DimensionInfo.NETHER ? LAVA_FLOW_MOTION_IN_NETHER : LAVA_FLOW_MOTION));
         }
 
         entity.addMotion(finalMotion);
@@ -322,7 +335,7 @@ public class AllayEntityPhysicsService implements EntityPhysicsService {
         var momentumMz = motion.z() * slipperinessMultiplier * MOMENTUM_FACTOR;
 
         // 2. Complete Formulas
-        var velocityFactor = entity.isOnGround() ? GROUND_VELOCITY_FACTOR : AIR_VELOCITY_FACTOR;
+        var velocityFactor = entity.isOnGround() ? entity.getDragFactorOnGround() : entity.getDragFactorInAir();
         var acceleration = velocityFactor * movementFactor;
         if (entity.isOnGround()) {
             acceleration *= (float) (effectFactor * pow(DEFAULT_FRICTION / slipperinessMultiplier, 3));
@@ -334,7 +347,7 @@ public class AllayEntityPhysicsService implements EntityPhysicsService {
 
         // Skip sprint jump boost because this service does not handle player's movement
 
-        var newMy = (motion.y() - (entity.hasGravity() ? entity.getGravity() : 0f)) * DRAG_FACTOR;
+        var newMy = (motion.y() - (entity.hasGravity() ? entity.getGravity() : 0f)) * (1 - entity.getDragFactorInAir());
         entity.setMotion(checkMotionThreshold(new Vector3f(newMx, newMy, newMz)));
     }